pyproject.toml
[project]
name = "scale"
version = "0.1.0"
description = "Scaling machine learning decoders for quantum error correction"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "lightning>=2.5.2",
    "matplotlib>=3.10.3",
    "matplotlib-label-lines>=0.8.1",
    "numpy>=2.3.1",
    "omegaconf>=2.3.0",
    "pandas>=2.3.1",
    "pylibcugraphops-cu12>=24.12.0",
    "pymatching>=2.2.2",
    "scipy>=1.16.0",
    "stim>=1.15.0",
    "torch>=2.7.1",
    "torch-geometric>=2.6.1",
    "torchmetrics>=1.7.4",
    "tqdm>=4.67.1",
    "wandb>=0.21.0",
    "schedulefree>=1.2.6",
    "beliefmatching>=0.1.1",
    "ipykernel>=6.30.0",
    "transformers>=4.53.3",
    "qiskit>=2.1.1",
    "hydra-submitit-launcher>=1.2.0",
]
